fixed server rights

+++ app/Http/Controllers/ServerController.php
@@ -5,8 +5,11 @@ namespace App\Http\Controllers;
 use App\Restart;
 use App\Server;
 use App\ServerInfo;
+use App\ServerOwner;
 use App\Source;
+use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class ServerController extends Controller
 {
@@ -17,7 +20,16 @@ class ServerController extends Controller
 
     public function index()
     {
-        $servers = Server::all();
+        $tempServers = Server::all();
+        $servers = [];
+        foreach($tempServers as $server) {
+            if ($server->hasOwner(auth()->user()->id)) {
+                array_push($servers, $server);
+            }
+        }
+
+        if(auth()->user()->admin)
+            $servers = Server::all();
 
         return view("servers.index", compact("servers"));
     }
@@ -34,10 +46,12 @@ class ServerController extends Controller
     public function destroy($server_id)
     {
         $server = Server::findOrFail($server_id);
-        $sources = Source::where("server_id", "=", $server->id);
 
-        foreach($sources as $source)
-            $source->delete();
+        if(! $this->checkOwner($server))
+            return redirect()->back()->with("error", "No access to delete this server");
+
+        DB::table("sources")->where("server_id", "=", $server->id)->delete();
+
         $server->delete();
 
         return redirect()->back()->with("success", "Deleted server and its connected sources");
@@ -47,6 +61,11 @@ class ServerController extends Controller
     {
         $server = Server::findOrFail($server_id);
 
+        if(! $this->checkOwner($server))
+            return redirect()->back()->with("error", "No access to delete this server");
+
+        $users = User::all();
+
         $info = $server->info;
 
         if(empty($server->info)) {
@@ -54,13 +73,17 @@ class ServerController extends Controller
             $info->save();
         }
 
-        return view("server_info.edit", compact("info"));
+        return view("server_info.edit", compact("info", "users"));
     }
 
 
     public function updateServerInfo($server_info_id, Request $request)
     {
         $serverInfo = ServerInfo::findOrFail($server_info_id);
+        $server = Server::findOrFail($serverInfo->server_id);
+        if(! $this->checkOwner($server))
+            return redirect()->back()->with("error", "No access to delete this server");
+
         if($request->has("players"))
             $serverInfo->players = $request->get("players");
         if($request->has("max_play_time"))
@@ -68,5 +91,31 @@ class ServerController extends Controller
 
         $serverInfo->save();
         return redirect()->back()->with("success", "Succesfully updated server info");
+    }
+
+    public function setOwners($server_id, Request $request)
+    {
+        if(! auth()->user()->admin)
+            return redirect()->back()->with("error", "No permission");
+
+        DB::table("server_owners")->where("server_id", "=", $server_id)->delete();
+
+        $newOwners = $request->get("select-owner");
+
+        foreach($newOwners as $owner) {
+            $serverOwner = new ServerOwner(["server_id" => $server_id, "user_id" => $owner]);
+            $serverOwner->save();
+        }
+
+        return redirect()->back()->with("success", "Successfully updated server owners");
+    }
+
+    public function checkOwner($server)
+    {
+        if(auth()->user()->admin)
+            return true;
+        if(! $server->hasOwner(auth()->user()->id))
+            return false;
+        return true;
     }
 }
